+++ src/components/sample/sample.ts
@@ -1,16 +1,18 @@
 import './sample.scss';
-import { WebComponent } from '../../common/WebComponent';
+import WebComponent, { createElementFromString } from '../../common/WebComponent';
 
 class SampleComponent implements WebComponent {
-  render(): string {
-    return '<p class="my-p" id="myP">sample text</p>';
+  render(): Element {
+    return createElementFromString('<p class="my-p" id="myP">sample text</p>');
   }
+
   setup(): void {
     document.querySelector('#myP')!.addEventListener('click', this.toggleColor);
   }
+
   private toggleColor(evt: Event): void {
-    (evt.currentTarget as Element).classList.toggle('my-p');
+    (evt.currentTarget as HTMLElement).classList.toggle('my-p');
   }
 }
 
-export { SampleComponent };
+export default SampleComponent;