+++ src/index.ts
@@ -1,7 +1,20 @@
 import './style.scss';
-import { SampleComponent } from './components/sample/sample';
+import SampleComponent from './components/sample/sample';
+import Router from './common/Router';
+import { createElementFromString } from './common/WebComponent';
 
-const root = document.querySelector('#root')!;
-const x = new SampleComponent();
-root.innerHTML = x.render();
-x.setup();
+const router = new Router(document.querySelector('#root')!);
+
+if (router.checkPath('')) {
+  router.renderComponent(createElementFromString('<a href="/#/sample">Go to sample component</a>'));
+  router.addRefreshListener(document.querySelector('a')!, 'click');
+}
+
+if (router.checkPath('sample')) {
+  const sampleComponent = new SampleComponent();
+  router.renderComponent(sampleComponent.render());
+  sampleComponent.setup();
+
+  router.renderComponent(createElementFromString('<a href="/">Go back</a>'));
+  router.addRefreshListener(document.querySelector('a')!, 'click');
+}