+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - HTML Validator.
 - Web Component interface.
+- Router class.
 
 ## [0.1.1] - 2021-01-27